Use class to encapsulate Bulk Tagger functionality
